requirements.txt
requests==2.32.5
PyYAML==6.0.2
loguru
pydantic>=2.6
pymupdf
pytesseract
Pillow
langdetect
faiss-cpu
pytest
